Move legacy DB adapter to own file
